+++ tests/RuleKit.Tests/Models/DecimalDigitsModel.cs
@@ -1,0 +1,11 @@
+﻿namespace RuleKit.Tests.Models
+{
+    internal sealed class DecimalDigitsModel
+    {
+        /// <summary>
+        /// Gets or sets the value to validate.
+        /// </summary>
+        [DecimalDigits(3, 2)]
+        public decimal? Value { get; set; }
+    }
+}